feat: MapController clearMarkers/clearPolylines & make counter Long

+++ org.envirocar.map/src/main/java/org/envirocar/map/model/Marker.kt
@@ -14,7 +14,7 @@ import androidx.annotation.DrawableRes
  * @property drawable The drawable of the marker.
  */
 class Marker private constructor(
-    val id: Int,
+    val id: Long,
     val point: Point,
     val title: String?,
     @DrawableRes val drawable: Int?
@@ -47,6 +47,6 @@ class Marker private constructor(
         fun default(point: Point) = Builder(point).build()
 
         @Volatile
-        private var count = 0
+        private var count = 0L
     }
 }
